backend/sst-health-service/pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
load-dotenv = "^0.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.1.0"
black = "^25.1.0"
flake8 = "^7.1.1"
isort = "^6.0.0"
yamllint = "^1.35.1"
pylint = "^3.3.4"
mypy = "^1.14.1"
pytest = "^8.3.4"
